fix tvOnAir bug

+++ lib/data/models/tv_model.dart
@@ -42,9 +42,9 @@ class TvModel extends Equatable {
         originalLanguage: json['original_language'],
         originalName: json['original_name'],
         overview: json['overview'],
-        popularity: json['popularity'],
+        popularity: json['popularity'].toDouble(),
         posterPath: json['poster_path'],
-        voteAverage: json['vote_average'],
+        voteAverage: json['vote_average'].toDouble(),
         voteCount: json['vote_count'],
       );
